% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_ideal}
\alias{sim_ideal}
\title{Simulate an ideal population}
\usage{
sim_ideal(
  init_pop_size,
  vaccinations,
  cases_novac,
  ve,
  lag,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{init_pop_size}{Inital population size}

\item{vaccinations}{Number of vaccinations at every timepoint}

\item{cases_novac}{Number of cases at every timepoint}

\item{ve}{Vaccine effectiveness}

\item{lag}{Lag period measured in timepoints}

\item{seed}{Integer seed to use}
}
\value{
A \link[tibble]{tibble} with the following columns:
  \item{timepoint}{Index of timepoint}
  \item{vaccinations}{Expected number of vaccinations}
  \item{cases_novac}{Expected number of cases in absence of vaccination}
  \item{ve}{Expected vaccine effectiveness}
  \item{pflu}{Flu incidence}
  \item{cases}{Actual number of cases}
  \item{popn}{Non-cases in absence of vaccination}
  \item{pvac}{Proportion of starting population vaccinated}
  \item{b}{Number vaccinated at that time}
  \item{A_to_E}{Number moved from A to E at that time}
  \item{A}{Non-vaccinated non-cases}
  \item{B}{Vaccinated non-cases lagging}
  \item{E}{Non-vaccinated cases}
}
\description{
Simulates an ideal population using the reference model from Tokars (2018).
}
\references{
Tokars JI, Rolfes MA, Foppa IM, Reed C. An evaluation and update
  of methods for estimating the number of influenza cases averted by
  vaccination in the United States. Vaccine. 2018;36(48):7331–7337.
  doi:10.1016/j.vaccine.2018.10.026
}
