% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_ideal}
\alias{sim_ideal}
\title{Simulate an ideal population}
\usage{
sim_ideal(
  init_pop_size,
  vaccinations,
  cases_novac,
  ve,
  lag,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{init_pop_size}{Inital population size}

\item{vaccinations}{Number of vaccinations at every timepoint}

\item{cases_novac}{Number of cases at every timepoint}

\item{ve}{Vaccine effectiveness}

\item{lag}{Lag period measured in timepoints}

\item{seed}{Integer seed to use}
}
\value{
A \link[tibble]{tibble} with the following columns:
\item{timepoint}{Index of timepoint}
}
\description{
Simulate an ideal population
}
